added some libraries to the project
Getting a newer version of jquery into the site as well.  Should probably look at updating the included jquery, but that is for later.

Also have to get the theme for the research page looking proper.  It is very bear bones right now.

Missing a lot of the research content but when I finish the front end code the databases just have to be entered in.